Implement BinaryDuplicateProcessIDs on darwin
* Implement BinaryDuplicateProcessIDs using libproc API available in Mac
  OS 10.5+.
* Fix Path() on darwin to return a non-null-terminated Go string.
* Move non-platform-specific funcs FindDuplicateProcess and
  DuplicateProcessIDs from executable_linux.go to executable.go.
