+++ executable/executable.go
@@ -3,7 +3,9 @@
 package executable
 
 import (
+	"fmt"
 	"log"
+	"os"
 	"path/filepath"
 	"syscall"
 
@@ -40,4 +42,31 @@ func Dir() (dir string, err error) {
 	}
 	dir, _ = filepath.Split(path)
 	return
+}
+
+// FindDuplicateProcess looks for any other processes with the same
+// binary name as passed in and returns the first one found.
+func FindDuplicateProcess(binary string) (*os.Process, int, error) {
+	all, err := BinaryDuplicateProcessIDs(binary)
+	if err != nil {
+		return nil, 0, err
+	}
+	if len(all) > 0 {
+		p, err := os.FindProcess(all[0])
+		if err != nil {
+			return nil, 0, err
+		}
+		return p, all[0], nil
+	}
+	return nil, 0, nil
+}
+
+// DuplicateProcessIDs returns all pids belonging to processes with the
+// same binary name as the running program.
+func DuplicateProcessIDs() (pids []int, err error) {
+	binary, err := Path()
+	if err != nil {
+		return nil, fmt.Errorf("Can't get path: %v", err)
+	}
+	return BinaryDuplicateProcessIDs(binary)
 }
